Add model log convi utils `log_loss_metric(loss, epoch)` and `define_loss_plot()` with examples; rework reading of training logs to handle non-time-series plots

+++ rafiki/utils/log.py
@@ -31,18 +31,19 @@ class JobLogger():
         self._log_line(type='METRICS', **kwargs)
 
     # Clears all logs (excluding plot definitions) before a specific time
-    def clear_logs(self, datetimeBefore=None):
-        if datetimeBefore is None:
-            datetimeBefore = datetime.now()
+    def clear_logs(self, datetime_before=None):
+        if datetime_before is None:
+            datetime_before = datetime.now()
         
         self._log_file.seek(0)
         new_log_file = tempfile.NamedTemporaryFile(delete=False, mode='w+', encoding='utf-8')
 
-        # Only copy over lines in new file that are not before `datetimeBefore`
+        # Only copy over lines in new file that are not before `datetime_before`
         for line in self._log_file:
             (log_datetime, _) = self._parse_line(line)
-            log_datetime = datetime.strptime(log_datetime, JOB_LOGGER_DATETIME_FORMAT)
-            if log_datetime is None or log_datetime >= datetimeBefore:
+            log_datetime = datetime.strptime(log_datetime, JOB_LOGGER_DATETIME_FORMAT) \
+                            if log_datetime is not None else None
+            if log_datetime is None or log_datetime >= datetime_before:
                 new_log_file.write(line)
 
         # Switch to new log file
@@ -63,6 +64,7 @@ class JobLogger():
 
     # Import and completely replace all logs
     def import_logs(self, logs_bytes):
+        if logs_bytes is None: return
         self._log_file.seek(0)
         self._log_file.write(logs_bytes.decode('utf-8'))
         self._log_file.truncate()
@@ -70,18 +72,18 @@ class JobLogger():
     '''
     Read logs as (plots, metrics, messages)
 
-    plots: { title: Plot }
+    plots: Plot[]
     Plot: { title, metrics, x_axis }
-    metrics: { name: Metric }
-    Metric: { name, values: [(Datetime, value)] }
-    messages: [(Datetime, message)]
+    metrics: Metric[]
+    Metric: { time: Datetime, [name]: [value]}
+    messages: { time: Datetime, message: string }[]
     Datetime: string (%Y-%m-%dT%H:%M:%S)
     '''
     def read_logs(self):
         self._log_file.seek(0)
 
-        plots = {}
-        metrics = {}
+        plots = []
+        metrics = []
         messages = []
         for line in self._log_file:
             (log_datetime, log_dict) = self._parse_line(line)
@@ -93,20 +95,21 @@ class JobLogger():
             del log_dict['type']
 
             if log_type == 'MESSAGE':
-                messages.append((log_datetime, log_dict.get('message')))
+                messages.append({
+                    'time': log_datetime,
+                    'message': log_dict.get('message')
+                })
 
             elif log_type == 'METRICS':
-                for (name, value) in log_dict.items():
-                    if name not in metrics:
-                        metrics[name] = { 
-                            'name': name,
-                            'values': []
-                        }
-                    metrics[name]['values'].append((log_datetime, value))
+                metrics.append({
+                    'time': log_datetime,
+                    **log_dict 
+                })
 
             elif log_type == 'PLOT':
-                title = log_dict['title']
-                plots[title] = log_dict
+                plots.append({
+                    **log_dict
+                })
 
         return (plots, metrics, messages)
 
@@ -129,6 +132,7 @@ class JobLogger():
         log_datetime = None
         if 'time' in log:
             log_datetime = log['time']
+            del log['time']
 
         return (log_datetime, log)
 
@@ -157,12 +161,13 @@ def _test_job_logger_for_train_worker():
     (plots, metrics, messages) = l2.read_logs()
     l2.destroy()
 
-    assert len(plots) == 1 and 'Model Loss' in plots
-    assert plots['Model Loss'] == { 'title': 'Model Loss', 'metrics': ['loss'], 'x_axis': None }
-    assert len(metrics) == 2 and 'loss' in metrics and 'learning_rate' in metrics
-    assert [x[1] for x in metrics['loss'].get('values')] == [3.42, 3.21, 3.11]
-    assert [isinstance(x[0], str) for x in metrics['learning_rate'].get('values')] == [True, True] 
-    assert [x[1] for x in messages] == ['START', 'END']
+    assert len(plots) == 1
+    assert plots[0] == { 'title': 'Model Loss', 'metrics': ['loss'], 'x_axis': None }
+    assert len(metrics) == 3
+    assert [x.get('loss') for x in metrics] == [3.42, 3.21, 3.11]
+    assert [x.get('learning_rate') for x in metrics] == [0.01, 0.01, None]
+    assert [isinstance(x.get('time'), str) for x in metrics] == [True, True, True] 
+    assert [x.get('message') for x in messages] == ['START', 'END']
     
 def _test_job_logger_for_predictor():
     l = JobLogger()
@@ -191,10 +196,11 @@ def _test_job_logger_for_predictor():
     l2.import_logs(logs_bytes)
     (plots, metrics, messages) = l2.read_logs()
     l2.destroy()
-    assert len(plots) == 1 and 'Queries' in plots
-    assert len(metrics) == 1 and 'query' in metrics
-    assert len(metrics['query'].get('values')) == 23
-    assert [x[1] for x in messages] == ['UP']
+
+    assert len(plots) == 1
+    assert len(metrics) == 23
+    assert metrics[4].get('query') == True
+    assert [x.get('message') for x in messages] == ['UP']
 
     # Predictor receives more queries
     time.sleep(1)
@@ -220,10 +226,12 @@ def _test_job_logger_for_predictor():
     l2.import_logs(logs_bytes)
     (plots, metrics, messages) = l2.read_logs()
     l2.destroy()
-    assert len(plots) == 1 and 'Queries' in plots
-    assert len(metrics) == 1 and 'query' in metrics
-    assert len(metrics['query'].get('values')) == 43
-    assert [x[1] for x in messages] == ['KILLED', 'DOWN']
+
+    assert len(plots) == 1
+    assert len(metrics) == 43
+    assert metrics[8].get('query') == True
+    assert isinstance(metrics[9].get('time'), str) is True
+    assert [x.get('message') for x in messages] == ['KILLED', 'DOWN']
 
 if __name__ == '__main__':
     print('Testing `JobLogger` for train worker...')
